requirements.txt
azure-identity
azure-mgmt-resource
tenacity
dacite
typing-extensions
aiohttp
pytest
pytest-asyncio